requirements.txt
aiohttp>=3.8.0
tqdm>=4.65.0
python-magic>=0.4.27 